% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_droughtshift_object}
\alias{create_droughtshift_object}
\title{create_droughtshift_object}
\usage{
create_droughtshift_object(droughtshift_input, drought_years, termination_year)
}
\arguments{
\item{droughtshift_input}{A tibble with headers: water_year, precip, flow, status. The status column must contain columns "pre_drought", "drought" and can also contain "termination" and "post-drought". See example by running generate_example_droughtshift_input(). This column will be overwritten if drought_years and termination_year are specified.}

\item{drought_years}{A vector of water years (integer)}

\item{termination_year}{Single integer. Year of drought termination.}
}
\value{
a tibble of storage target levels by week
}
\description{
takes annual time totals and converts to droughtshift object for plotting and robustness analysis
}
