% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droughtshift-package.R
\docType{package}
\name{droughtshift-package}
\alias{droughtshift}
\alias{droughtshift-package}
\title{droughtshift: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
