% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{generate_example_droughtshift_input}
\alias{generate_example_droughtshift_input}
\title{generate_example_droughtshift_input}
\usage{
generate_example_droughtshift_input()
}
\value{
an example input dataset for droughtshift
}
\description{
this function generates an example input dataset for droughtshift
}
