% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_droughtshift_object}
\alias{plot_droughtshift_object}
\title{create_droughtshift_object}
\usage{
plot_droughtshift_object(droughtshift_object)
}
\arguments{
\item{droughtshift_object}{output from \code{create_droughtshift_object()}}
}
\value{
a ggplot object
}
\description{
plots a droughtshift object
}
